requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
sqlalchemy[asyncio]>=2.0.27
asyncpg>=0.29.0
pydantic>=2.6.0
pydantic-settings>=2.6.0
email-validator>=2.1.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.9
alembic>=1.13.0
pytz>=2024.1
python-dotenv>=1.0.0
